Refactor modules pruning to filter and latent_estimation to count
