+++ cleanlab/__init__.py
@@ -1,9 +1,9 @@
 from .version import __version__
 from . import classification
 from . import latent_algebra
-from . import latent_estimation
+from . import count
 from . import noise_generation
 from . import polyplex
-from . import pruning
+from . import filter
 from . import util
 from . import baseline_methods
