+++ tests/test_latent_estimation_and_pruning.py
@@ -17,8 +17,8 @@
 from __future__ import (
     print_function, absolute_import, division, unicode_literals,
     with_statement, )
-from cleanlab import latent_estimation
-from cleanlab import pruning
+from cleanlab import count
+from cleanlab import filter
 from cleanlab.latent_algebra import compute_inv_noise_matrix
 from cleanlab.noise_generation import generate_noise_matrix_from_trace
 from cleanlab.noise_generation import generate_noisy_labels
@@ -115,7 +115,7 @@ data = make_data(sparse=False, seed=1)
 def test_exact_prune_count():
     remove = 5
     s = data['s']
-    noise_idx = pruning.get_noise_indices(
+    noise_idx = filter.get_noise_indices(
         s=s,
         psx=data['psx'],
         num_to_remove_per_class=remove,
@@ -128,21 +128,21 @@ def test_exact_prune_count():
 def test_pruning_both(n_jobs):
     remove = 5
     s = data['s']
-    class_idx = pruning.get_noise_indices(
+    class_idx = filter.get_noise_indices(
         s=s,
         psx=data['psx'],
         num_to_remove_per_class=remove,
         prune_method='prune_by_class',
         n_jobs=n_jobs,
     )
-    nr_idx = pruning.get_noise_indices(
+    nr_idx = filter.get_noise_indices(
         s=s,
         psx=data['psx'],
         num_to_remove_per_class=remove,
         prune_method='prune_by_noise_rate',
         n_jobs=n_jobs,
     )
-    both_idx = pruning.get_noise_indices(
+    both_idx = filter.get_noise_indices(
         s=s,
         psx=data['psx'],
         num_to_remove_per_class=remove,
@@ -365,7 +365,7 @@ def test_get_noise_indices_multi_label():
     s_ml = [[z, data['y_train'][i]] for i, z in enumerate(data['s'])]
     for multi_label in [True, False]:
         for prune_method in ['prune_by_class', 'prune_by_noise_rate']:
-            noise_idx = pruning.get_noise_indices(
+            noise_idx = filter.get_noise_indices(
                 s=s_ml if multi_label else data['s'],
                 psx=data['psx'],
                 prune_method=prune_method,
